Unity compatibility patch / .NET 2 + 2 Subset
This set of patches allow Backtrace to be compiled for Unity in .NET 2/.NET 2 Subset mode. We have also added #ifdefs for compatibility with versions prior to Unity 2017.4.

Key changes:
- Removed nameof(...) operator, and swapped to strings.
- Changed NET35 to (NET35 || NET_2_0 || NET_2_0_SUBSET) to match corresponding Unity auto-defines. (sorry did this with search/replace, it's a little messy -- too many to do manually in any reasonable time period)
- Converted $"" to String.Format
- Removed method?. and replaced with if/then/else or ternary operator

Special notes have been added with "// TODO" comments for sections which should be checked and tested more robustly.

Mostly untested.