xbox native client: fixes output directory for minidumps
Uses the updated API of backtrace_native_xbox.dll, previously
backtrace_native_xbox_mt.dll, to specify a path for writing dump files.
Prior to this commit, it was unpredictable as to whether a minidump could
be written or not as the current directory of the process was
(unintentionally) being used.

With this change, minidumps are written to the value specified by the
configuration's database path is passed which, by default, is
`Application.persistentDataPath`.

Removed unused exports, BacktraceCrash, used only for debugging, and
BacktraceSetUrl, which was folded into BacktraceNativeXboxInit.

The change will require updating to the latest backtrace_native_xbox.dll.

[Internal Ref BT-509]