metrics: fix for pending jobs being delayed one or more frames
The list of pending jobs was being traversed using an index that was
unconditionally incremented, even after removing an item from the list.
This could cause a job to be skipped in `SendPendingEvents`, at least
until the next frame.

This was not a fatal error as any skipped jobs would be handled in the
later calls to SendPendingEvents (next tick).

Noticed while auditing the code for a reported issue with dropped events.