metrics: retry submission on connection failure
Previously, only a subset protocol errors would be considered for retry
when submitting metrics (summed or unique) to events.backtrace.io. This
excluded, notably, a failure to connect to the server at all. If the
device was offline at startup, for instance, then the
"Application Launches" event would never get sent.

This commit changes the handling so that any network error will also
go through the retry logic for submitting an event.