Expose BacktraceMetrics.SessionId and BacktraceClient.AttributeProvider
This is useful to allow SessionId and other attributes (e.g. 'guid') to be
written to 'dying memory' on platforms that support it (e.g. Switch,
PlayStation), which the Backtrace server is then able to parse on importing
from platform crash reporting services, and set on the imported errors. It
can then be correlated with other errors from the same session.